Hidden pages

Pages can be hidden with the hide function.

List of hidden pages

Hidden page 1

Hidden page 2

Hidden page 3

Absolute paths

Hidden 1

First heading

Second heading

Hidden 2

Hidden 3

Relative paths

Equivalent to what I have in lib/contents.md.

Relative paths

Just have the index.md in to show it does parse this block.

Hidden pages

List of hidden pages

Docs for hide

Docs for hide

Documenter.hide — Function.

hide(page)

Allows a page to be hidden in the navigation menu. It will only show up if it happens to be the current page. The hidden page will still be present in the linear page list that can be accessed via the previous and next page links. The title of the hidden page can be overriden using the => operator as usual.

Usage

makedocs(
    ...,
    pages = [
        ...,
        hide("page1.md"),
        hide("Title" => "page2.md")
    ]
)

source

hide(root, children)

Allows a subsection of pages to be hidden from the navigation menu. root will be linked to in the navigation menu, with the title determined as usual. children should be a list of pages (note that it can not be hierarchical).

Usage

makedocs(
    ...,
    pages = [
        ...,
        hide("Hidden section" => "hidden_index.md", [
            "hidden1.md",
            "Hidden 2" => "hidden2.md"
        ]),
        hide("hidden_index.md", [...])
    ]
)

source