DocMeta

DocMeta

Documenter.DocMeta.initdocmeta! — Function.

source

Documenter.DocMeta.META — Constant.

The unique Symbol that is used to store the metadata dictionary in each module.

source

Documenter.DocMeta.METAMODULES — Constant.

List of modules that have the metadata dictionary added.

source

Documenter.DocMeta.METATYPE — Type.

Type of the metadata dictionary.

source

Documenter.DocMeta.VALIDMETA — Constant.

Dictionary of all valid metadata keys and their types.

source