DocumenterTools

DocumenterTools

DocumenterTools.package_devpath — Function.

package_devpath(pkg)

Returns the path to the top level directory of a devved out package source tree. The package is identified by its top level module pkg.

Generator

DocumenterTools.Generator — Module.

Provides the functions related to generating documentation stubs.

DocumenterTools.Generator.gitignore — Method.

gitignore()

Contents of the default .gitignore file.

DocumenterTools.Generator.index — Method.

index(pkgname)

Contents of the default src/index.md file.

DocumenterTools.Generator.make — Method.

make(pkgname; format)

Contents of the default make.jl file.

DocumenterTools.Generator.mkdocs — Method.

mkdocs(pkgname; description, author, url)

Contents of the default mkdocs.yml file.

DocumenterTools.Generator.project — Method.

project(; format)

Contents of the default Project.toml file.

DocumenterTools.Generator.savefile — Method.

savefile(f, root, filename)

Attempts to save a file at $(root)/$(filename). f will be called with file stream (see open).

filename can also be a file in a subdirectory (e.g. src/index.md), and then then subdirectories will be created automatically.