Expanders

Documenter.Expanders — Module.

Defines node "expanders" that transform nodes from the parsed markdown files.

source

Documenter.Expanders.AutoDocsBlocks — Type.

Parses each code block where the language is @autodocs and replaces it with all the docstrings that match the provided key/value pairs Modules = ... and Order = ....

```@autodocs
Modules = [Foo, Bar]
Order   = [:function, :type]
```

source

Documenter.Expanders.ContentsBlocks — Type.

Parses each code block where the language is @contents and replaces it with a nested list of all Header nodes in the generated document. The pages and depth of the list can be set using Pages = [...] and Depth = N where N is and integer.

```@contents
Pages = ["foo.md", "bar.md"]
Depth = 1
```

The default Depth value is 2.

source

Documenter.Expanders.DocsBlocks — Type.

Parses each code block where the language is @docs and evaluates the expressions found within the block. Replaces the block with the docstrings associated with each expression.

```@docs
Documenter
makedocs
deploydocs
```

source

Documenter.Expanders.EvalBlocks — Type.

Parses each code block where the language is @eval and evaluates it's content. Replaces the block with the value resulting from the evaluation. This can be useful for inserting generated content into a document such as plots.

```@eval
using PyPlot
x = linspace(-π, π)
y = sin(x)
plot(x, y, color = "red")
savefig("plot.svg")
Markdown.parse("![Plot](plot.svg)")
```

source

Documenter.Expanders.ExampleBlocks — Type.

Parses each code block where the language is @example and evaluates the parsed Julia code found within. The resulting value is then inserted into the final document after the source code.

```@example
a = 1
b = 2
a + b
```

source

Documenter.Expanders.ExpanderPipeline — Type.

The default node expander "pipeline", which consists of the following expanders:

TrackHeaders

MetaBlocks

DocsBlocks

AutoDocsBlocks

EvalBlocks

IndexBlocks

ContentsBlocks

ExampleBlocks

REPLBlocks

source

Documenter.Expanders.IndexBlocks — Type.

Parses each code block where the language is @index and replaces it with an index of all docstrings spliced into the document. The pages that are included can be set using a key/value pair Pages = [...] such as

```@index
Pages = ["foo.md", "bar.md"]
```

source

Documenter.Expanders.MetaBlocks — Type.

Parses each code block where the language is @meta and evaluates the key/value pairs found within the block, i.e.

```@meta
CurrentModule = Documenter
DocTestSetup  = quote
    using Documenter
end
```

source

Documenter.Expanders.REPLBlocks — Type.

Similar to the ExampleBlocks expander, but inserts a Julia REPL prompt before each toplevel expression in the final document.

source

Documenter.Expanders.TrackHeaders — Type.

Tracks all Markdown.Header nodes found in the parsed markdown files and stores an Anchors.Anchor object for each one.

source