Writers

Documenter.Writers — Module.

Provides a rendering function, render, for writing each supported Formats.Format to file.

Note that currently Formats.Markdown is the only supported format.

source

Documenter.Writers.Writer — Type.

A parametric type that allows us to use multiple dispatch to pick the appropriate writer for each output format.

The parameter f should be an instance of the Formats.Format enumeration.

source

Documenter.Writers.render — Method.

Writes a Documents.Document object to .user.build directory in the format specified in .user.format.

The method should be overloaded in each writer as

render(::Writer{format}, doc)

where format is one of the values of the Formats.Format enumeration.

source

Documenter.Writers.MarkdownWriter — Module.

Provides the render methods to write the documentation as Markdown files (MIME"text/plain").

source

Documenter.Writers.HTMLWriter — Module.

Provides the render methods to write the documentation as HTML files (MIME"text/html").

source

Documenter.Writers.HTMLWriter.HTMLContext — Type.

HTMLWriter-specific globals that are passed to domify and other recursive functions.

source

Documenter.Writers.HTMLWriter.collect_subsections — Method.

Returns a list of tuples (anchor, text), corresponding to all level 2 headers.

source

Documenter.Writers.HTMLWriter.copy_asset — Method.

Copies an asset from Documenters assets/html/ directory to doc.user.build. Returns the path of the copied asset relative to .build.

source

Documenter.Writers.HTMLWriter.domify — Method.

Converts recursively a Documents.Page, Base.Markdown or Documenter *Node objects into HTML DOM.

source

Documenter.Writers.HTMLWriter.getpage — Method.

Returns a page (as a Documents.Page object) using the HTMLContext.

source

Documenter.Writers.HTMLWriter.mdconvert — Method.

Convert a markdown object to a DOM.Node object.

The parent argument is passed to allow for context-dependant conversions.

source

Documenter.Writers.HTMLWriter.navhref — Method.

Get the relative hyperlink between two Documents.NavNodes. Assumes that both Documents.NavNodes have an associated Documents.Page (i.e. .page is not null).

source

Documenter.Writers.HTMLWriter.navitem — Method.

navitem returns the lists and list items of the navigation menu. It gets called recursively to construct the whole tree.

source

Documenter.Writers.HTMLWriter.pagetitle — Method.

Tries to guess the page title by looking at the <h1> headers and returns the header contents as a Nullable (nulled if the algorithm was unable to determine the header).

source

Documenter.Writers.HTMLWriter.relhref — Method.

Calculates a relative HTML link from one path to another.

source

Documenter.Writers.HTMLWriter.render_page — Method.

Constructs and writes the page referred to by the navnode to .build.

source

Documenter.Writers.LaTeXWriter — Module.

Provides the render methods to write the documentation as LaTeX files (MIME"text/latex").

source