DocsXRefTests

On the page (unlike in the g docstring) can link directly to AbstractSelector because the CurrentModule is Main.

API

Main.DocsReferencingMain.f — Function

This is the function f.

It references Documenter.Selectors.AbstractSelector, which resolves only because of a fallback to Main.

See also g.

Main.DocsReferencingMain.g — Function

This is the function g

If references Main.AbstractSelector, which should resolve, unlike the non-fully-qualified AbstractSelector (even though AbstractSelector is in Main)

See also f.

Documenter.Selectors.AbstractSelector — Type

Root selector type. Each user-defined selector must subtype from this, i.e.

abstract type MySelector <: Selectors.AbstractSelector end

abstract type First  <: MySelector end
abstract type Second <: MySelector end

Documenter.hide — Function

hide(page)

Allows a page to be hidden in the navigation menu. It will only show up if it happens to be the current page. The hidden page will still be present in the linear page list that can be accessed via the previous and next page links. The title of the hidden page can be overridden using the => operator as usual.

Usage

makedocs(
    ...,
    pages = [
        ...,
        hide("page1.md"),
        hide("Title" => "page2.md")
    ]
)

hide(root, children)

Allows a subsection of pages to be hidden from the navigation menu. root will be linked to in the navigation menu, with the title determined as usual. children should be a list of pages (note that it can not be hierarchical).

Usage

makedocs(
    ...,
    pages = [
        ...,
        hide("Hidden section" => "hidden_index.md", [
            "hidden1.md",
            "Hidden 2" => "hidden2.md"
        ]),
        hide("hidden_index.md", [...])
    ]
)