Anchors

Documenter.Anchors — Module.

Defines the Anchor and AnchorMap types.

Anchors and AnchorMaps are used to represent links between objects within a document.

source

Documenter.Anchors.Anchor — Type.

Stores an arbitrary object called .object and it's location within a document.

Fields

object – the stored object.

order – ordering of object within the entire document.

file – the destination file, in build, where the object will be written to.

id – the generated "slug" identifying the object.

nth – integer that unique-ifies anchors with the same id.

source

Documenter.Anchors.AnchorMap — Type.

Tree structure representating anchors in a document and their relationships with eachother.

Object Hierarchy

id -> file -> anchors

Each id maps to a file which in turn maps to a vector of Anchor objects.

source

Documenter.Anchors.add! — Method.

add!(m, anchor, id, file)

Adds a new Anchor to the AnchorMap for a given id and file.

Either an actual Anchor object may be provided or any other object which is automatically wrapped in an Anchor before being added to the AnchorMap.

source

Documenter.Anchors.anchor — Method.

anchor(m, id)

Returns the Anchor object matching id. file and n may also be provided. A Nullable{Anchor} is returned which must be unwrapped with isnull and get before use.

source

Documenter.Anchors.exists — Method.

exists(m, id, file, n)

Does the given id exist within the AnchorMap? A file and integer n may also be provided to narrow the search for existance.

source

Documenter.Anchors.isunique — Method.

isunique(m, id)

Is the id unique within the given AnchorMap? May also specify the file.

source