Anchors

Documenter.Anchor — Type

Stores an arbitrary object called .object and it's location within a document.

Fields

object – the stored object.

order – ordering of object within the entire document.

file – the destination file, in build, where the object will be written to.

id – the generated "slug" identifying the object.

nth – integer that unique-ifies anchors with the same id.

Documenter.AnchorMap — Type

Tree structure representing anchors in a document and their relationships with each other.

Object Hierarchy

id -> file -> anchors

Each id maps to a file which in turn maps to a vector of Anchor objects.

Documenter.anchor — Method

anchor(m, id)

Returns the Anchor object matching id. file and n may also be provided. An Anchor is returned, or nothing in case of no match.

Documenter.anchor_add! — Method

anchor_add!(m, anchor, id, file)

Adds a new Anchor to the AnchorMap for a given id and file.

Either an actual Anchor object may be provided or any other object which is automatically wrapped in an Anchor before being added to the AnchorMap.

Documenter.anchor_exists — Method

anchor_exists(m, id, file, n)

Does the given id exist within the AnchorMap? A file and integer n may also be provided to narrow the search for existence.

Documenter.anchor_fragment — Method

Create an HTML fragment from an anchor.

Documenter.anchor_isunique — Method

anchor_isunique(m, id)

Is the id unique within the given AnchorMap? May also specify the file.

Documenter.anchor_label — Method

Create a label from an anchor.