Builder

Documenter.Builder — Module

Defines the Documenter.jl build "pipeline" named DocumentPipeline.

Each stage of the pipeline performs an action on a Documents.Document object. These actions may involve creating directory structures, expanding templates, running doctests, etc.

Documenter.Builder.CheckDocument — Type

Checks that all documented objects are included in the document and runs doctests on all valid Julia code blocks.

Documenter.Builder.CrossReferences — Type

Finds and sets URLs for each @ref link in the document to the correct destinations.

Documenter.Builder.Doctest — Type

Runs all the doctests in all docstrings and Markdown files.

Documenter.Builder.DocumentPipeline — Type

The default document processing "pipeline", which consists of the following actions:

SetupBuildDirectory

Doctest

ExpandTemplates

CrossReferences

CheckDocument

Populate

RenderDocument

Documenter.Builder.ExpandTemplates — Type

Executes a sequence of actions on each node of the parsed markdown files in turn.

Documenter.Builder.Populate — Type

Populates the ContentsNodes and IndexNodes with links.

Documenter.Builder.RenderDocument — Type

Writes the document tree to the build directory.

Documenter.Builder.SetupBuildDirectory — Type

Creates the correct directory layout within the build folder and parses markdown files.

Documenter.Builder.walk_navpages — Method

walk_navpages(visible, title, src, children, parent, doc)

Recursively walks through the Documents.Document's .user.pages field, generating Documents.NavNodes and related data structures in the process.

This implementation is the de facto specification for the .user.pages field.