Builder

Documenter.Builder — Module.

Defines the Documenter.jl build "pipeline" named DocumentPipeline.

Each stage of the pipeline performs an action on a Documents.Document object. These actions may involve creating directory structures, expanding templates, running doctests, etc.

source

Documenter.Builder.CheckDocument — Type.

Checks that all documented objects are included in the document and runs doctests on all valid Julia code blocks.

source

Documenter.Builder.CrossReferences — Type.

Finds and sets URLs for each @ref link in the document to the correct destinations.

source

Documenter.Builder.DocumentPipeline — Type.

The default document processing "pipeline", which consists of the following actions:

SetupBuildDirectory

ExpandTemplates

CrossReferences

CheckDocument

Populate

RenderDocument

source

Documenter.Builder.ExpandTemplates — Type.

Executes a sequence of actions on each node of the parsed markdown files in turn.

source

Documenter.Builder.Populate — Type.

Populates the ContentsNodes and IndexNodes with links.

source

Documenter.Builder.RenderDocument — Type.

Writes the document tree to the build directory.

source

Documenter.Builder.SetupBuildDirectory — Type.

Creates the correct directory layout within the build folder and parses markdown files.

source

Documenter.Builder.walk_navpages — Method.

walk_navpages(ps, parent, doc)

Recursively walks through the Documents.Document's .user.pages field, generating Documents.NavNodes and related data structures in the process.

This implementation is the de facto specification for the .user.pages field.

source