CrossReferences

Documenter.CrossReferences — Module

Provides the crossref function used to automatically calculate link URLs.

Documenter.CrossReferences.crossref — Method

crossref(doc)

Traverses a Documents.Document and replaces links containg @ref URLs with their real URLs.

Documenter.CrossReferences.find_object — Method

find_object(doc, binding, typesig)

Find the included Object in the doc matching binding and typesig. The matching heuristic isn't too picky about what matches and will only fail when no Bindings matching binding have been included.