DocChecks

Documenter.DocChecks — Module.

Provides two functions, missingdocs and doctest, for checking docs.

source

Documenter.DocChecks.doctest — Method.

doctest(doc)

Traverses the document tree and tries to run each Julia code block encountered. Will abort the document generation when an error is thrown. Use doctest = false keyword in Documenter.makedocs to disable doctesting.

source

Documenter.DocChecks.missingdocs — Method.

missingdocs(doc)

Checks that a Documents.Document contains all available docstrings that are defined in the modules keyword passed to Documenter.makedocs.

Prints out the name of each object that has not had its docs spliced into the document.

source