DocMeta

Documenter.DocMeta.initdocmeta! — Function

Documenter.DocMeta.META — Constant

The unique Symbol that is used to store the metadata dictionary in each module.

Documenter.DocMeta.METAMODULES — Constant

List of modules that have the metadata dictionary added.

Documenter.DocMeta.METATYPE — Type

Type of the metadata dictionary.

Documenter.DocMeta.VALIDMETA — Constant

Dictionary of all valid metadata keys and their types.