DocSystem

Documenter.DocSystem — Module

Provides a consistent interface to retreiving DocStr objects from the Julia docsystem in both 0.4 and 0.5.

Documenter.DocSystem.binding — Method

Converts an object to a Base.Docs.Binding object.

binding(any)

Supported inputs are:

Binding

DataType

Function

Module

Symbol

Note that unsupported objects will throw an ArgumentError.

Documenter.DocSystem.convertmeta — Method

convertmeta(meta)

Converts a 0.4-style docstring cache into a 0.5 one.

The original docstring cache is not modified.

Documenter.DocSystem.docstr — Method

docstr(md; kws...)

Construct a DocStr object from a Markdown.MD object.

The optional keyword arguments are used to add new data to the DocStr's .data dictionary.

Documenter.DocSystem.getdocs — Function

getdocs(object)
getdocs(object, typesig; kws...)

Accepts objects of any type and tries to convert them to Bindings before searching for the Binding in the docsystem.

Note that when conversion fails this method returns an empty Vector{DocStr}.

Documenter.DocSystem.getdocs — Function

getdocs(binding)
getdocs(binding, typesig; compare, modules, aliases)

Find all DocStr objects that match the provided arguments:

binding: the name of the object.

typesig: the signature of the object. Default: Union{}.

compare: how to compare signatures? Exact (==) or subtypes (<:). Default: <:.

modules: which modules to search through. Default: all modules.

aliases: check aliases of binding when nothing is found. Default: true.

Returns a Vector{DocStr} ordered by definition order in 0.5 and by type_morespecific in 0.4.

Documenter.DocSystem.multidoc — Function

Construct a MultiDoc object from the provided argument.

Valid inputs are:

Markdown.MD

Docs.FuncDoc

Docs.TypeDoc

Documenter.DocSystem.parsedoc — Method

DocSystem.parsedoc(docstr::DocStr)

Thin internal wrapper around Base.Docs.parsedoc which prints additional debug information in case Base.Docs.parsedoc fails with an exception.