DocTests

DocTests

Documenter.DocTests — Module.

Provides the doctest function that makes sure that the jldoctest code blocks in the documents and docstrings run and are up to date.

source

Documenter.DocTests.doctest — Method.

doctest(doc)

Traverses the document tree and tries to run each Julia code block encountered. Will abort the document generation when an error is thrown. Use doctest = false keyword in Documenter.makedocs to disable doctesting.

source