Documenter

Documenter.gitrm_copy — Function

gitrm_copy(src, dst)

Uses git rm -r to remove dst and then copies src to dst. Assumes that the working directory is within the git repository of dst is when the function is called.

This is to get around #507 on filesystems that are case-insensitive (e.g. on OS X, Windows). Without doing a git rm first, git add -A will not detect case changes in filenames.

Documenter.git_push — Function

git_push(
    root, tmp, repo;
    branch="gh-pages", dirname="", target="site", sha="", devurl="dev",
    deploy_config, folder,
)

Handles pushing changes to the remote documentation branch. The documentation are placed in the folder specified by subfolder.

Documenter.user_host_upstream — Function

user_host_upstream(repo)

Disassemble repo address into user, host, and path to repo. If no user is given, default to "git". Reassemble user, host and path into an upstream to git push to.