Documents

Documents

Documenter.Documents — Module.

Defines Document and its supporting types

Page

User

Internal

Globals

source

Documenter.Documents.Document — Type.

Represents an entire document.

source

Documenter.Documents.Globals — Type.

Page-local values such as current module that are shared between nodes in a page.

source

Documenter.Documents.Internal — Type.

Private state used to control the generation process.

source

Documenter.Documents.NavNode — Type.

Element in the navigation tree of a document, containing navigation references to other page, reference to the Page object etc.

source

Documenter.Documents.Page — Type.

Represents a single markdown file.

source

Documenter.Documents.User — Type.

User-specified values used to control the generation process.

source

Documenter.Documents.navpath — Method.

Constructs a list of the ancestors of the navnode (inclding the navnode itself), ordered so that the root of the navigation tree is the first and navnode itself is the last item.

source

Documenter.Documents.populate! — Method.

populate!(document)

Populates the ContentsNodes and IndexNodes of the document with links.

This can only be done after all the blocks have been expanded (and nodes constructed), because the items have to exist before we can gather the links to those items.

source

Documenter.Documents.walk — Method.

walk(f, meta, element)

Calls f on element and any of its child elements. meta is a Dict containing metadata such as current module.

source