Documents

Documenter.Document — Type

Represents an entire document.

Documenter.Globals — Type

Page-local values such as current module that are shared between nodes in a page.

Documenter.Internal — Type

Private state used to control the generation process.

Documenter.NavNode — Type

Element in the navigation tree of a document, containing navigation references to other page, reference to the Page object etc.

Documenter.Page — Type

Represents a single markdown file.

Documenter.User — Type

User-specified values used to control the generation process.

Documenter.getplugin — Method

getplugin(doc::Document, T)

Retrieves the Plugin type for T stored in doc. If T was passed to makedocs, the passed type will be returned. Otherwise, a new T object will be created using the default constructor T().

Documenter.navpath — Method

Constructs a list of the ancestors of the navnode (including the navnode itself), ordered so that the root of the navigation tree is the first and navnode itself is the last item.

Documenter.populate! — Method

populate!(document)

Populates the ContentsNodes and IndexNodes of the document with links.

This can only be done after all the blocks have been expanded (and nodes constructed), because the items have to exist before we can gather the links to those items.

Documenter.relpath_from_remote_root — Method

relpath_from_remote_root(remotes, path)

Returns the the the remote that contains the file, and the relative path of the file within the repo (or nothing, nothing if the file is not in a known repo).

Documenter.shortcommit — Method

shortcommit(remoteref)

Returns the first 5 characters of the current Git commit hash of the remote.