Documents

Documenter.Document — Type

Represents an entire document.

Documenter.Globals — Type

Page-local values such as current module that are shared between nodes in a page.

Documenter.Internal — Type

Private state used to control the generation process.

Documenter.LocalImage — Type

Represents a reference to a local image. The path can be assumed to be an "absolute" path relative to the document root (i.e. src/ or build/ directories). See LocalLink for more details.

Documenter.LocalLink — Type

Represents a reference to a local file. The path can be assumed to be an "absolute" path relative to the document root (i.e. src/ or build/ directories).

In the standard setup, when the documentation setup lives in docs/, with source files in docs/src, a link to the file docs/src/foo/bar.md would have path = "foo/bar.md".

Documenter.NavNode — Type

Element in the navigation tree of a document, containing navigation references to other page, reference to the Page object etc.

Documenter.Page — Type

Represents a single markdown file.

Documenter.User — Type

User-specified values used to control the generation process.

Documenter.getplugin — Method

Documenter.getplugin(doc::Document, T) -> Plugin

Retrieves the object for the Plugin sub-type T stored in doc. If an object of type T was an element of the plugins list passed to makedocs, that object will be returned. Otherwise, a new T object will be created using the default constructor T(). Subsequent calls to getplugin(doc, T) return the same object.

Documenter.is_strict — Method

is_strict(::Document, val::Symbol) -> Bool

Internal function to check if Documenter should throw an error or simply print a warning when hitting error condition.

Single-argument is_strict(strict) provides a curried function.

Documenter.navpath — Method

Constructs a list of the ancestors of the navnode (including the navnode itself), ordered so that the root of the navigation tree is the first and navnode itself is the last item.

Documenter.populate! — Method

populate!(document)

Populates the ContentsNodes and IndexNodes of the document with links.

This can only be done after all the blocks have been expanded (and nodes constructed), because the items have to exist before we can gather the links to those items.

Documenter.relpath_from_remote_root — Method

relpath_from_remote_root(doc, path)

Returns the the the remote that contains the file, and the relative path of the file within the repo (or nothing, nothing if the file is not in a known repo).

Documenter.shortcommit — Method

shortcommit(remoteref)

Returns the first 5 characters of the current Git commit hash of the remote.