Walkers

Walkers

Documenter.Walkers — Module.

Provides the walk function.

source

Documenter.Walkers.walk — Method.

walk(f, meta, element)

Calls f on element and any of its child elements. meta is a Dict containing metadata such as current module.

source