Writers

Writers

Documenter.Writers — Module.

A module that provides several renderers for Document objects. The supported formats are currently:

:markdown – the default format.

:html – generates a complete HTML site with navigation and search included.

:latex – generates a PDF using LuaLaTeX.

source

Documenter.Writers.render — Method.

Writes a Documents.Document object to .user.build directory in the formats specified in the .user.format vector.

Adding additional formats requires adding new Selector definitions as follows:

abstract type CustomFormat <: FormatSelector end

Selectors.order(::Type{CustomFormat}) = 4.0 # or a higher number.
Selectors.matcher(::Type{CustomFormat}, fmt, _) = fmt === :custom
Selectors.runner(::Type{CustomFormat}, _, doc) = CustomWriter.render(doc)

# Definition of `CustomWriter` module below...

source

Documenter.Writers.MarkdownWriter — Module.

A module for rendering Document objects to markdown.

source

Documenter.Writers.HTMLWriter — Module.

A module for rendering Document objects to HTML.

Keywords

HTMLWriter uses the following additional keyword arguments that can be passed to Documenter.makedocs: authors, pages, sitename, version. The behavior of HTMLWriter can be further customized by setting the format keyword of Documenter.makedocs to a HTML, which accepts the following keyword arguments: analytics, assets, canonical, disable_git, edit_branch and prettyurls.

sitename is the site's title displayed in the title bar and at the top of the *navigation menu. This argument is mandatory for HTMLWriter.

pages defines the hierarchy of the navigation menu.

Experimental keywords

version specifies the version string of the current version which will be the selected option in the version selector. If this is left empty (default) the version selector will be hidden. The special value git-commit sets the value in the output to git:{commit}, where {commit} is the first few characters of the current commit hash.

HTML Plugin options

The HTML Documenter.Plugin provides additional customization options for the HTMLWriter. For more information, see the HTML documentation.

Page outline

The HTMLWriter makes use of the page outline that is determined by the headings. It is assumed that if the very first block of a page is a level 1 heading, then it is intended as the page title. This has two consequences:

It is then used to automatically determine the page title in the navigation menu and in the <title> tag, unless specified in the .pages option.

If the first heading is interpreted as being the page title, it is not displayed in the navigation sidebar.

source

Documenter.Writers.HTMLWriter.HTML — Type.

HTML(kwargs...)

Sets the behavior of HTMLWriter.

Keyword arguments

prettyurls (default true) – allows toggling the pretty URLs feature.

By default (i.e. when prettyurls is set to true), Documenter creates a directory structure that hides the .html suffixes from the URLs (e.g. by default src/foo.md becomes src/foo/index.html, but can be accessed with via src/foo/ in the browser). This structure is preferred when publishing the generate HTML files as a website (e.g. on GitHub Pages), which is Documenter's primary use case.

If prettyurls = false, then Documenter generates src/foo.html instead, suitable for local documentation builds, as browsers do not normally resolve foo/ to foo/index.html for local files.

To have pretty URLs disabled in local builds, but still have them enabled for the automatic CI deployment builds, you can set prettyurls = get(ENV, "CI", nothing) == "true" (the specific environment variable you will need to check may depend on the CI system you are using, but this will work on Travis CI).

disable_git can be used to disable calls to git when the document is not in a Git-controlled repository. Without setting this to true, Documenter will throw an error and exit if any of the Git commands fail. The calls to Git are mainly used to gather information about the current commit hash and file paths, necessary for constructing the links to the remote repository.

edit_branch specifies which branch, tag or commit the "Edit on GitHub" links point to. It defaults to master. If it set to nothing, the current commit will be used.

canonical specifies the canonical URL for your documentation. We recommend you set this to the base url of your stable documentation, e.g. https://juliadocs.github.io/Documenter.jl/stable. This allows search engines to know which version to send their users to. See wikipedia for more information. Default is nothing, in which case no canonical link is set.

analytics can be used specify the Google Analytics tracking ID.

assets can be used to include additional assets (JS, CSS, ICO etc. files). See below for more information.

Default and custom assets

Documenter copies all files under the source directory (e.g. /docs/src/) over to the compiled site. It also copies a set of default assets from /assets/html/ to the site's assets/ directory, unless the user already had a file with the same name, in which case the user's files overrides the Documenter's file. This could, in principle, be used for customizing the site's style and scripting.

The HTML output also links certain custom assets to the generated HTML documents, specifically a logo and additional javascript files. The asset files that should be linked must be placed in assets/, under the source directory (e.g /docs/src/assets) and must be on the top level (i.e. files in the subdirectories of assets/ are not linked).

For the logo, Documenter checks for the existence of assets/logo.{svg,png,webp,gif,jpg,jpeg}, in this order. The first one it finds gets displayed at the top of the navigation sidebar.

Additional JS, ICO, and CSS assets can be included in the generated pages using the assets keyword for makedocs. assets must be a Vector{String} and will include each listed asset in the <head> of every page in the order in which they are listed. The type of the asset (i.e. whether it is going to be included with a <script> or a <link> tag) is determined by the file's extension – either .js, .ico, or .css. Adding an ICO asset is primarilly useful for setting a custom favicon.

source

Documenter.Writers.HTMLWriter.MDBlockContext — Constant.

MDBlockContext is a union of all the Markdown nodes whose children should be blocks. It can be used to dispatch on all the block-context nodes at once.

source

Documenter.Writers.HTMLWriter.HTMLContext — Type.

HTMLWriter-specific globals that are passed to domify and other recursive functions.

source

Documenter.Writers.HTMLWriter.collect_subsections — Method.

Returns an ordered list of tuples, (toplevel, anchor, text), corresponding to level 1 and 2 headings on the page. Note that if the first header on the page is a level 1 header then it is not included – it is assumed to be the page title and so does not need to be included in the navigation menu twice.

source

Documenter.Writers.HTMLWriter.copy_asset — Method.

Copies an asset from Documenters assets/html/ directory to doc.user.build. Returns the path of the copied asset relative to .build.

source

Documenter.Writers.HTMLWriter.domify — Method.

Converts recursively a Documents.Page, Markdown or Documenter *Node objects into HTML DOM.

source

Documenter.Writers.HTMLWriter.fixlinks! — Method.

Replaces URLs in Markdown.Link elements (if they point to a local .md page) with the actual URLs.

source

Documenter.Writers.HTMLWriter.get_url — Method.

Returns the full path corresponding to a path of a .md page file. The the input and output paths are assumed to be relative to src/.

source

Documenter.Writers.HTMLWriter.get_url — Method.

Returns the full path of a Documents.NavNode relative to src/.

source

Documenter.Writers.HTMLWriter.getpage — Method.

Returns a page (as a Documents.Page object) using the HTMLContext.

source

Documenter.Writers.HTMLWriter.mdconvert — Method.

Convert a markdown object to a DOM.Node object.

The parent argument is passed to allow for context-dependant conversions.

source

Documenter.Writers.HTMLWriter.navhref — Method.

Get the relative hyperlink between two Documents.NavNodes. Assumes that both Documents.NavNodes have an associated Documents.Page (i.e. .page is not nothing).

source

Documenter.Writers.HTMLWriter.navitem — Method.

navitem returns the lists and list items of the navigation menu. It gets called recursively to construct the whole tree.

It always returns a DOM.Node. If there's nothing to display (e.g. the node is set to be invisible), it returns an empty text node (DOM.Node("")).

source

Documenter.Writers.HTMLWriter.open_output — Method.

Opens the output file of the navnode in write node. If necessary, the path to the output file is created before opening the file.

source

Documenter.Writers.HTMLWriter.pagetitle — Method.

Tries to guess the page title by looking at the <h1> headers and returns the header contents of the first <h1> on a page (or nothing if the algorithm was unable to find any <h1> headers).

source

Documenter.Writers.HTMLWriter.pretty_url — Method.

If prettyurls for HTML is enabled, returns a "pretty" version of the path which can then be used in links in the resulting HTML file.

source

Documenter.Writers.HTMLWriter.relhref — Method.

Calculates a relative HTML link from one path to another.

source

Documenter.Writers.HTMLWriter.render_page — Method.

Constructs and writes the page referred to by the navnode to .build.

source

Documenter.Writers.LaTeXWriter — Module.

A module for rendering Document objects to LaTeX and PDF.

Keywords

LaTeXWriter uses the following additional keyword arguments that can be passed to makedocs: authors, sitename.

sitename is the site's title displayed in the title bar and at the top of the navigation menu. It goes into the \title LaTeX command.

authors can be used to specify the authors of. It goes into the \author LaTeX command.

source

Documenter.Writers.LaTeXWriter.LaTeX — Type.

LaTeXWriter.LaTeX(; kwargs...)

Output format specifier that results in LaTeX/PDF output. Used together with makedocs, e.g.

makedocs(
    format = LaTeX()
)

The makedocs argument sitename will be used for the \title field in the tex document, and if the build is for a release tag (i.e. when the "TRAVIS_TAG" environment variable is set) the version number will be appended to the title. The makedocs argument authors should also be specified, it will be used for the \authors field in the tex document.

Keyword arguments

platform sets the platform where the tex-file is compiled, either "native" (default) or "docker". See Other Output Formats for more information.

source

Documenter.Plugin — Type.

abstract type Plugin end

Any plugin that needs to either solicit user input or store information in a Documents.Document should create a subtype of Plugin. The subtype, T <: Documenter.Plugin, must have an empty constructor T() that initialized T with the appropriate default values.

To retrieve the values stored in T, the plugin can call Documents.getplugin. If T was passed to makedocs, the passed type will be returned. Otherwise, a new T object will be created.

source